+++ test/upstream/extract_tests.ml
@@ -533,7 +533,8 @@ let variant_prefixes cls =
 let unreplayable ~defs ~config ~theme_vars ~classes expected =
   let routed =
     List.filter
-      (Tw_tools.Entrypoint.is_custom_routed ~defs:[] ~udefs:defs)
+      (Tw_tools.Entrypoint.is_custom_routed ~theme:Tw.Scheme.default ~defs:[]
+         ~udefs:defs)
       classes
   in
   let unreadable =